ProjectileHit entity hit support, target caster setting, target Projectile instead of Location, ProjectileListener improvements
